Update Vertex.java
I made the vertices comparable because that will come in handy for ordering vertices in edges and triangles
I also made it so that vertices with the same x and y value will be considered equal and return the same hashcode. I'm not sure if that is strictly necessary but oh well.