Create Triangle.java
Triangle contains three vertices. v[0] is the smallest vertex. v[1] is the next vertex in counterclockwise order and v[2] is the last vertex. it also has a insideCircumcircle function which is supposed to come in handy later. This function requires that the vertices in the triangle are in counter clock wise order. Similar to the edge class this class will also return true for the equals and the hashCode function as long as they have the same vertices regardless of the order the triangle is created in.